Move ast analysis to a separate file
