+++ src/fal/cli/model_generator/__init__.py
@@ -1,0 +1,1 @@
+from .model_generator import generate_python_dbt_models
